+++ Assets/Scripts/SpecialObjects/PassiveObject/NuclearExplosionModule/NuclearExplosionModule.cs
@@ -3,10 +3,10 @@ using UnityEngine;
 
 public class NuclearExplosionModule : PassiveObjects
 {
-    [SerializeField] private int explosionDamageMultiplier = 0;
+    [SerializeField] private float explosionDamageMultiplier = 0;
     [SerializeField] public static float nuclearDotTimer;
-    [SerializeField] public static int nuclearDotDamage;
-    [SerializeField] private int nuclearDotD = 0;
+    [SerializeField] public static float nuclearDotDamage;
+    [SerializeField] private float nuclearDotD = 0;
     [SerializeField] private float nuclearDotT = 0f;
